Consolidate BaseCompiler test files into a single comprehensive file
This commit consolidates all BaseCompiler test files into a single comprehensive file:
- base-compiler-tests.ts (original)
- base-compiler-tests-new.ts (new compilation tests)
- base-compiler-exec-tests.ts (execution tests)
- base-compiler-utils-tests.ts (utility tests)

The tests are organized into logical sections with clear headers, making it easier
to understand and maintain the BaseCompiler tests.

This addresses cross-platform compatibility by ensuring Windows-specific path handling
is properly taken into account in relevant tests.
